+++ tests/test_sync.py
@@ -97,6 +97,24 @@ def test_retry_inactive():
     assert 1 == num_called
 
 
+def test_retry_inactive_ok():
+    """
+    If inactive, the happy path still works.
+    """
+    num_called = 0
+
+    @stamina.retry(on=Exception)
+    def f():
+        nonlocal num_called
+        num_called += 1
+
+    stamina.set_active(False)
+
+    f()
+
+    assert 1 == num_called
+
+
 def test_retry_block():
     """
     Sync retry_context blocks are retried.
@@ -106,6 +124,9 @@ def test_retry_block():
     for attempt in stamina.retry_context(on=ValueError, wait_max=0):
         with attempt:
             i += 1
+
+            assert i == attempt.num
+
             if i < 2:
                 raise ValueError